chore: refactor and docs
