--- assets/cdp.yaml
@@ -1,14 +1,0 @@
-# editor to use when opening a project with the -e flag
-editor: nvim
-multiplexer: tmux
-
-source:
-  # Directories containing the project in your HOME
-  dirs:
-    - Repos
-  # Markers to identify the project root
-  project_markers:
-    - .git
-    - node_modules
-    - go.mod
-#
