+++ cmd/shell.go
@@ -6,11 +6,10 @@ import (
 	"slices"
 	"strings"
 
-	"github.com/alexpetrean80/cdp/executable"
+	"github.com/alexpetrean80/cdp/lib/executable"
 	"github.com/spf13/cobra"
 )
 
-// shellCmd represents the shell command
 var (
 	shellExecPath string
 	shell         executable.Program
@@ -25,7 +24,10 @@ var (
 			}
 
 			if i := slices.Index([]string{"bash", "sh", "zsh", "fish"}, shellExecPath); i == -1 {
-				return fmt.Errorf("%s is not a supported shell. valid options are sh, bash, zsh and fish", shellExecPath)
+				return fmt.Errorf(
+					"%s is not a supported shell. valid options are sh, bash, zsh and fish",
+					shellExecPath,
+				)
 			}
 
 			shell = executable.New(shellExecPath, args...)
@@ -37,5 +39,6 @@ var (
 
 func init() {
 	rootCmd.AddCommand(shellCmd)
-	shellCmd.Flags().StringVarP(&shellExecPath, "shell", "s", "", "shell to be opened (defaults to $SHELL)")
+	shellCmd.Flags().
+		StringVarP(&shellExecPath, "shell", "s", "", "shell to be opened (defaults to $SHELL)")
 }
