--- finder/finder_test.go
@@ -1,1 +1,0 @@
-package finder_test
