+++ lib/finder/finder.go
@@ -10,6 +10,9 @@ import (
 	"golang.org/x/sync/errgroup"
 )
 
+// Finder is a struct that contains the root directory to start the search from,
+// the markers to search for, the result channel to send the results to, and the
+// errgroup to manage the goroutines.
 type Finder struct {
 	RootDir string
 	Markers map[string]struct{}
@@ -17,6 +20,7 @@ type Finder struct {
 	Group   *errgroup.Group
 }
 
+// New creates a new Finder struct with the given root directory, markers, result
 func New(rootDir string, markers []string, resCh chan string, g *errgroup.Group) *Finder {
 	pf := Finder{}
 	pf.RootDir = rootDir
@@ -32,10 +36,15 @@ func New(rootDir string, markers []string, resCh chan string, g *errgroup.Group)
 	return &pf
 }
 
+// Find starts the search from the root directory and sends the results to the result channel.
 func (pf Finder) Find(name string) error {
 	return pf.findRec(pf.RootDir, name)
 }
 
+// findRec is a recursive function that searches for the given name in the root directory.
+// If the name is found, it sends the result to the result channel.
+// If the entry is a directory, it starts a new goroutine to search in the directory.
+// If the entry is a hidden directory, it skips the directory.
 func (pf Finder) findRec(rootDir, name string) error {
 	entries, err := os.ReadDir(rootDir)
 	if err != nil {
