+++ lib/lib.go
@@ -6,13 +6,16 @@ import (
 	"log"
 	"os"
 
-	"github.com/alexpetrean80/cdp/finder"
+	"github.com/alexpetrean80/cdp/lib/finder"
 	"github.com/ktr0731/go-fuzzyfinder"
 	"github.com/mitchellh/go-homedir"
 	"github.com/spf13/viper"
 	"golang.org/x/sync/errgroup"
 )
 
+// GetFullPathOfDirs is a function that returns the full path of the directories
+// specified in the configuration file.
+// It assumes that the directories are relative to the home directory.
 func GetFullPathOfDirs() ([]string, error) {
 	homeDir, err := homedir.Dir()
 	if err != nil {
@@ -27,6 +30,8 @@ func GetFullPathOfDirs() ([]string, error) {
 	return res, nil
 }
 
+// ChangeDirectory is a function that changes the current working directory to the
+// project directory specified by the name.
 func ChangeDirectory(name string, last bool) error {
 	projectPath, err := GetProjectPath(name, last)
 	if err != nil {
@@ -36,6 +41,7 @@ func ChangeDirectory(name string, last bool) error {
 	return os.Chdir(projectPath)
 }
 
+// ReadLastProject is a function that reads the last project from the $HOME/.local/share/cdp_last file.
 func ReadLastProject() (string, error) {
 	file, err := os.Open(fmt.Sprintf("%s/.local/share/cdp_last", os.Getenv("HOME")))
 	if err != nil {
@@ -49,6 +55,7 @@ func ReadLastProject() (string, error) {
 	return string(proj), nil
 }
 
+// WriteLastProject is a function that writes the last project to the $HOME/.local/share/cdp_last file.
 func WriteLastProject(projectPath string) error {
 	file, err := os.Create(fmt.Sprintf("%s/.local/share/cdp_last", os.Getenv("HOME")))
 	if err != nil {
@@ -63,6 +70,7 @@ func WriteLastProject(projectPath string) error {
 	return nil
 }
 
+// GetProjectPath is a function that returns the project path retrieved from finder.
 func GetProjectPath(name string, last bool) (string, error) {
 	if last {
 		projectPath, err := ReadLastProject()
@@ -105,6 +113,9 @@ func GetProjectPath(name string, last bool) (string, error) {
 	return projectPath, nil
 }
 
+// FindProjects is a function that searches for the projects with a filter.
+// If the filter is empty, it returns all the projects found in the directories
+// specified in the configuration file.
 func FindProjects(name string) <-chan string {
 	g := new(errgroup.Group)
 	ch := make(chan string, 10)
@@ -118,8 +129,6 @@ func FindProjects(name string) <-chan string {
 		if !isDirectory(dir) {
 			break
 		}
-		fmt.Println("dir:", dir)
-		fmt.Println("name:", name)
 
 		pf := finder.New(dir, markers, ch, g)
 		g.Go(func(rootDir string) func() error {
